Adapt to sx changes
FossilOrigin-Name: 182d16f4e292b077aa39716dc81e9d55e61b8f5c108042062ffae9c31158d922

+++ shtml/shtml.go
@@ -792,7 +792,7 @@ func (te *TransformEnv) flattenText(sb *strings.Builder, lst *sx.Pair) {
 
 type transformFn func([]sx.Object) sx.Object
 
-func (te *TransformEnv) bind(name string, minArity int, fn transformFn) {
+func (te *TransformEnv) bind(name string, minArity int16, fn transformFn) {
 	te.astEnv.Bind(te.astSF.MustMake(name), &sxeval.Builtin{
 		Name:     name,
 		MinArity: minArity,
